requirements.txt
pika==1.2.0
fastapi==0.72.0
uvicorn[standard]==0.17.0
pymongo==4.0.1
pydantic==1.9.0
python-decouple==3.5